requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
numpy>=1.26.0
opencv-python>=4.9.0
Pillow>=10.0.0
cryptography>=46.0.0
reportlab>=4.0.0
